server/requirements.txt
# FastAPI and server dependencies
fastapi==0.103.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0

# Database dependencies
motor==3.3.1
pymongo==4.5.0

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.3.0

# Machine Learning dependencies
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
transformers==4.33.2
scikit-learn==1.3.0

# Data processing dependencies
numpy==1.24.4
pandas==2.0.3
opencv-python==4.8.0.76
Pillow==10.0.0

# Visualization dependencies
matplotlib==3.7.2
seaborn==0.12.2

# Dataset downloading dependencies
datasets==2.14.4
kaggle==1.5.16

# Additional NLP dependencies
nltk==3.8.1
spacy==3.6.1

# Utility dependencies
tqdm==4.66.1
requests==2.31.0